chore: move ReaderService to proper Services namespace
