+++ LotComWatcher/Worker.cs
@@ -1,6 +1,7 @@
 using LotComWatcher.Models.Datasources;
 using LotComWatcher.Models.Datatypes;
 using LotComWatcher.Models.Enums;
+using LotComWatcher.Models.Services;
 
 namespace LotComWatcher;
